pyproject.toml
[project]
name = "cslr-exp-platform"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "datamodel-code-generator[http]>=0.53.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "typer>=0.15.0",
    "pydantic>=2.10.0",
    "rich>=13.0.0",
]

[tool.setuptools]
packages = ["ws_src"]

[project.scripts]
ws = "ws_src.commands:main"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "black>=26.1.0",
    "flake8>=7.3.0",
]
